+++ src/port/trace/consensus/ProposalAcceptRequestReceived.java
@@ -9,12 +9,12 @@ import util.trace.TraceableInfo;
 public class ProposalAcceptRequestReceived extends TraceableInfo {
 	
 
-	public ProposalAcceptRequestReceived(String aMessage, Object aSource, String anObjectName, int aProposalNumber, Object aProposal) {
+	public ProposalAcceptRequestReceived(String aMessage, Object aSource, String anObjectName, double aProposalNumber, Object aProposal) {
 		super(aMessage, aSource );
 	}
 	
 	
-	public static ProposalAcceptRequestReceived newCase(Object aSource, String anObjectName, int aProposalNumber, Object aProposal) {
+	public static ProposalAcceptRequestReceived newCase(Object aSource, String anObjectName, double aProposalNumber, Object aProposal) {
     	String aMessage =  anObjectName + "." + aProposalNumber + "=" + aProposal;
     	ProposalAcceptRequestReceived retVal = new ProposalAcceptRequestReceived(aMessage, aSource, anObjectName, aProposalNumber, aProposal);
    	    retVal.announce();
