+++ src/port/trace/consensus/ProposalAcceptRequestSent.java
@@ -9,12 +9,12 @@ import util.trace.TraceableInfo;
 public class ProposalAcceptRequestSent extends TraceableInfo {
 	
 
-	public ProposalAcceptRequestSent(String aMessage, Object aSource, String anObjectName, int aProposalNumber, Object aProposal) {
+	public ProposalAcceptRequestSent(String aMessage, Object aSource, String anObjectName, double aProposalNumber, Object aProposal) {
 		super(aMessage, aSource );
 	}
 	
 	
-	public static ProposalAcceptRequestSent newCase(Object aSource, String anObjectName, int aProposalNumber, Object aProposal) {
+	public static ProposalAcceptRequestSent newCase(Object aSource, String anObjectName, double aProposalNumber, Object aProposal) {
     	String aMessage =  anObjectName + "." + aProposalNumber + "=" + aProposal;
     	ProposalAcceptRequestSent retVal = new ProposalAcceptRequestSent(aMessage, aSource, anObjectName, aProposalNumber, aProposal);
    	    retVal.announce();
