+++ src/port/trace/consensus/ProposalWaitStarted.java
@@ -9,12 +9,12 @@ import util.trace.TraceableInfo;
 public class ProposalWaitStarted extends TraceableInfo {
 	
 
-	public ProposalWaitStarted(String aMessage, Object aSource, String anObjectName, int aProposalNumber, Object aProposal) {
+	public ProposalWaitStarted(String aMessage, Object aSource, String anObjectName, double aProposalNumber, Object aProposal) {
 		super(aMessage, aSource );
 	}
 	
 	
-	public static ProposalWaitStarted newCase(Object aSource, String anObjectName, int aProposalNumber, Object aProposal) {
+	public static ProposalWaitStarted newCase(Object aSource, String anObjectName, double aProposalNumber, Object aProposal) {
     	String aMessage =  anObjectName + "." + aProposalNumber + "=" + aProposal;
     	ProposalWaitStarted retVal = new ProposalWaitStarted(aMessage, aSource, anObjectName, aProposalNumber, aProposal);
    	    retVal.announce();
